Ensure all model instances have camelcase copies of all attributes
Javascript attributes are in camelcase by convention, so this allows this
library to fit in better in JS, while maintaining some degree of backward
compatibility. However, the iconicTaxonName() method on Taxon had to be
renamed to getIconicTaxonName() in order to not conflict with the camelcase
version of the iconic_taxon_name attributes, necessitating a major version
change.